refactor(nika-check): descend native-first under the 15k wall
crate-size: move native-first classification into
nika-check-analyzer so nika-check production LOC drops below
15000. Historical classify/classify_all stay as forwarding
wrappers at nika_check::native_first.

fn-length: extract instantiate_skeleton, expand_query, and
classify_explicit_done so run / route_impl / classify_turn
sit at <=100.

C03: TTY vs JSON for mcp:spotify/search lives in nika-display
(check_render hardcodes NIKA-BUILTIN-001). JSON findings
already stamp NIKA-INVOKE-001. CLI lane owns the renderer.
